modified xml-reader

+++ eti-cmdline/devices/xml-filereader/xml-descriptor.h
@@ -65,7 +65,7 @@ public:
 	             	                 std::string	byteOrder);
 	void		addChannelOrder (int channelOrder,
 	                                 std::string Value);
-	void		add_dataBlock (int currBlock,  int Count,
+	void		add_dataBlock (int currBlock,  uint64_t Count,
                                        int  blockNumber, std::string Unit);
 	void		add_freqtoBlock	(int blockno, int freq);
 	void		add_modtoBlock (int blockno, std::string modType);
