+++ eti-cmdline/devices/xml-filereader/xml-reader.cpp
@@ -51,7 +51,7 @@ struct timeval tv;
 
 	xml_Reader::xml_Reader (FILE		*f,
 	                        xmlDescriptor	*fd,
-	                        uint32_t	filePointer,
+	                        uint64_t	filePointer,
 	                        RingBuffer<std::complex<float>> *b,
 	                        bool	continue_on_eof) {
 	this	-> file		= f;
@@ -94,7 +94,7 @@ void	xml_Reader::stopReader	() {
 
 static	int cycleCount = 0;
 void	xml_Reader::run () {
-int	samplesRead	= 0;
+uint64_t	samplesRead	= 0;
 uint64_t	nextStop;
 int	startPoint	= filePointer;
 
@@ -139,9 +139,9 @@ int	startPoint	= filePointer;
 	}
 }
 
-int	xml_Reader::compute_nrSamples (FILE *f, int blockNumber) {
-int	nrElements	= fd -> blockList. at (blockNumber). nrElements;
-int	samplesToRead	= 0;
+uint64_t	xml_Reader::compute_nrSamples (FILE *f, int blockNumber) {
+uint64_t	nrElements	= fd -> blockList. at (blockNumber). nrElements;
+uint64_t	samplesToRead	= 0;
 
 	(void)f;
 	if (fd -> blockList. at (blockNumber). typeofUnit == "Channel") {
@@ -159,7 +159,7 @@ int	samplesToRead	= 0;
 	return samplesToRead;
 }
 
-int	xml_Reader::readSamples (FILE *theFile, 
+uint64_t	xml_Reader::readSamples (FILE *theFile, 
 	                         void(xml_Reader::*r)(FILE *theFile,
 	                                    std::complex<float> *, int)) {
 std::complex<float> temp [2048];
